Bicubic Interpolation Resize

+++ pixel.h
@@ -10,10 +10,10 @@ class Pixel
 {
 public:
 
-    uint8_t red;
-    uint8_t green;
-    uint8_t blue;
-    uint8_t alpha; 
+    int16_t red;
+    int16_t green;
+    int16_t blue;
+    int16_t alpha; 
 
     ~Pixel() = default;
     Pixel(const Pixel& other) = default;
